Update base container image
The alpine image used by `linkerd-proxy-init` includes some vulnerable
packages. Even though these packages are not actively used, it's worth
updating the base image to a more recent version.

This change also updates to `golang` base image to from 1.16.2 to
1.16.9-alpine3.14. This way we build everything on the same platform
that it runs on and also minimize the size of the container image.