Add iptables-nft support
This change introduces support for iptables-nft. To support backwards
compatibility, iptables-nft usage is optional and off by default.

This PR introduces a new flag "use-nft-backend". When the flag is set to
'true', it will call into iptables-nft binary instead of iptables. The
binary itself will be present in the alpine container, and is added when
installing the iptables package.

Instead of harcoding the binary name in all commands, we now pass the
binary name as an argument.

Signed-off-by: Matei David <matei@buoyant.io>