Fix issue with arbitrary iptables binary
To support `nft` we have added support to use arbitrary iptables
binaries. Prior to this change, in order to use the `--wait` flag for
iptables, we'd check that the command path would be equal to "iptables".

After the change, we check if the path has "iptables" as a substring.
This has the unfortunate side-effect of applying the `-w` flag to any
command that contains the word "iptables", including "iptables-save"
which does not support the command.

This causes the container to fail when the -w flag is specified. The CNI
plugin uses the iptables library and the -w flag and this issue will
block it from working.

Signed-off-by: Matei David <matei@buoyant.io>