Refactor CI pipeline (#67)
This PR refactors the set-up and build process to be more consistent
with the tools and processed that are used in the other Linkerd repositories.

Changes made as part of this:

Replace make with just as a build tool
Add a workflow to lint existing actions
Overhaul release workflow (replace outdated actions, build for multi-arch)
Change image tags in tests to use cr.l5d.io instead of ghcr. Being
consistent here helps with manifests that have been injected where
only the tag needs to be changed instead of the registry.
Instead of a makefile, we now have a justfile. Instead of copying recipes,
we now have only a few recipes that aid with building or testing locally.
The recipes are used only the integration-tests workflow, since that would
mirror the set-up in CI.

We no longer build the images in the test-runner script. The test runner itself
should (imo) be separate from the build process, even if we are only building
a test image.

* Run integration-tests in k3d
* Add Action Lint workflow
* Add justfile
* Fix errors in CI workflows
* Add just to integration test workflow


Signed-off-by: Matei David <matei@buoyant.io>
Co-authored-by: Alejandro Pedraza <alejandro@buoyant.io>