Add sys net caps to xtables-nft-multi
Both iptables and iptables-nft use the same packet matching code in the
kernel (xtables). iptables binaries are symbolic links to the respective
xtables binaries; for example, iptables-legacy would be a symbolic link
to xtables-legacy-multi, and likewise, iptables-nft a symbolic link to
xtables-nft-multi.

To support running the container as non-root in Kubernetes, instead of
adding system capabilities in the pod spec, they have been baked into
the image; xtables-legacy-multi has been granted both net_admin and
net_raw capabilities. To preserve the same behaviour when using
nftables, this change adds the same capabilities to the
xtables-nft-multi binary.

Signed-off-by: Matei David <matei@buoyant.io>