Add Docker Cache Prune to just recipe
In the release workflow, we're calling an inexistent
'bin/docker-prune-cache' script. The script will delete all lingering
files in the cache that are no longer in use or referenced by the cache
manifest file (thereby cleaning up storage).

This change extracts the logic from the script (present in the linkerd2
repo) and adds it as a justfile recipe, to be called in the release
workflow.

This change also formats the justfile through `just --fmt`

Signed-off-by: Matei David <matei@buoyant.io>