Pods may be initialized without any Linkerd specific routing rules when the CNI plugin is used. This typically tends to happen when a node is rebooted (or drained). Pods can be initialized on the node (host) before the CNI plugin has had a chance to configure the plugin list, missing out on iptables commands.
This change introduces a simple container to test whether a pod has been
configured with the required routing rules. To test for the existence of
routing rules, the container starts a server bound on the proxy's
outbound port (default: 4140). After the server is started, the
container will attempt a TCP connection to the k8s api server; if
outbound redirection rules exist in the pod's network namespace, then
the connection should be established to the server listening on :4140.

Under normal conditions, when rules are set-up, no outbound connections
should be successful (since they would require passing through the
proxy, which would not be listening when using an init container). In an
effort to have more of a comprehensive check, we additionally generate
some random data. The server will send the bytes to the client, if the
bytes are received successfully, then validation passes.

Any connection error (including timeouts), or a non-valid random
sequence will result in the container exiting with code 111 (errno 111:
Connection Refused). This will allow a yet-to-be-implemented controller
to restart pods with errors.
Move shared logic into pkg and proxy-init in own dir

Signed-off-by: Matei David <matei@buoyant.io>