Add validator to build process (#103)
This change adds the validator to the repository's build process.
Instead of maintaing a separate Dockerfile, and a separate release step
in the workflow (along with additional just recipes, and so on), the
validator is built in the same Dockerfile as proxy-init. The resulting
binary is added to the runtime image.

The validator can be run by modifying the entrypoint in the manifest to
call `linkerd-network-validator` instead of proxy-init. The validator is
statically linked to musl.

Signed-off-by: Matei David <matei@buoyant.io>