ci: Move `just action-prune-docker` to a script
Currently the `justfile` invokes `cargo` to determine the validator
version. This means that it will install the Rust version in
`rust-toolchain` when the runner OS includes another Rust version. This
is unnecessary in the proxy-init release workflow.

This change removes the `action-prune-docker` just recipe to a dedicated
script in `.github/docker-prune.sh` so that it can be run without any of
the `justfile` dependencies. This also makes the script covered by `just
sh-lint`.

Signed-off-by: Oliver Gould <ver@buoyant.io>