Add Calico scenario (#182)
Support for calico in integration tests. As part of this change, the cni-plugin test harness has been modified to support a more generic pattern for scenario creation. Additionally, a test-util has been introduced for the CNI plugin. Along with these two changes, the justfile and workflow file had to be changed to support the new tests.

Signed-off-by: Matei David <matei@buoyant.io>