Introduce a running mode for CNI plugin
Currently, the CNI plugin will adapt to its environment dynamically,
depending on the presence of a configuration file. For example, when a
configuration file is present, it simply appends itself to it.
Otherwise, it will create its own configuration file.

In certain scenarios, such as when a conflist file is deleted, inotify
mgiht not be quick enough to capture the change, and as such will
attempt to install linkerd-cni in interface mode. After this, it will
attempt to delete its own configuration file, but since it has already
been observed by inotify, it will compute the diff hash and think it has
to be re-created, thereby getting into an endless loop.

fixes #linkerd2/10375

Signed-off-by: Matei David <matei@buoyant.io>