feat(linkerd-cni): add support for plain iptables commands
Currently the `iptables-mode` for linkerd-cni admits the values `legacy`
and `default`, which make the plugin use the `iptables-legacy[-save]`
and `iptables-nft[-save]` commands respectively.

This assumes those commands are available in the node environment, given
that linkerd-cni is triggered by the kubelet.

We have found that not to be the case for RHEL, where by default only
`iptables[-save]` is available, which is equivalent to the
`iptables-nft[-save]` command in other enviroments.

To address this case, this change adds a new possible value
`iptables-mode: default` that makes the plugin use the `iptables[-save]`
commands.

This has been tested successfully using RKE2 deployed in RHEL 8.10.