build(repair-controller): remove openssl dependency (#553)
Removed openssl dependencies, in favor of rustls + aws-lc-rs.

This implied updating the repair-controller's main.rs to enable the default aws-lc provider.

Also the following dependencies were upgraded:

- k8s-openapi from 0.20 to 0.25
- kube from 0.87 to 1.1 (this implied some updates to repair-controller's lib.rs)
- kubert from 0.21 to 0.25
- prometheus-client from 0.22 to 0.23
- Added an explicit dependency to rustls 0.23
- golang from 1.23 to 1.24 in the Dockerfile-tester

Finally, the repair-controller integration test was unnecessarily building the controller in-line, even if it's built withing the cni-plugin image itself, so this got simplified.