fix(cni-plugin): remove square brackets from host which may not be ipv6 in kubeconfig (#589)
When the install-cni script generates a kubeconfig for the cni-plugin, it wraps the kubernetes service host in square brackets.  This is not valid if the host is an ipv4 address and, as of recent versions of Go, such hosts will fail to be parsed by the `net/url` package.  See: https://github.com/golang/go/commit/d6d2f7bf76718f1db05461cd912ae5e30d7b77ea

Remove the square brackets so that this host can be parsed properly.

Signed-off-by: Alex Leong <alex@buoyant.io>